dependencies: revert from python-systemd to systemd-python for f22
Fedora 23 renamed its systemd Python packages to python{,3}-systemd to
conform to Python packaging guidelines.  F22 and older don't have the new
names available so revert back to the old name in our rpm spec.

The new packages have a virtual provide for the old name so that dependency
works on both versions.