journalpump: Add support for elasticsearch as an output
With this change we can also send the logs directly to an elasticsearch
instance to make them immediately searchable.

Also make max_send_interval configurable, we may want to do the same
for the minimum batch size. (currently 100)

Especially in the case of Elasticsearch you do not want to send
batches that are too small since it is much less efficient than
sending large batches to it.