Stop reading journal when in-memory queue grows too large
When there was an issue sending out log messages the system previously
read an unbounded amount of messages into memory, resulting in the app
causing OOM errors if the situation persisted for long enough.

Now the max buffer size parameter is actually honored and readers that
have any senders with more than the configured max outbound messages
pending are removed from the list of journals being polled. They are
returned to the poll list once the send buffers are again below
threshold.