reader: if set, only pass through messages that match searches
If searches have been defined for a given reader, only pass through
messages that match those searches to senders. This allows one to
output a selection of the captured messages to readers.

Previous behavior of passing full incoming stream and doing stats
on searches can be accomplished by configuring two readers the
searches on the first and the senders on the latter.