journalpump: clean up function interfaces
* remove unused function get_resume_cursor

This function is called from nowhere in the existing code and can be
removed.

* make create_journald_reader_if_missing private

This method is not used except for internal JournalReader accounting and
is made private to more clearly define the JournalReader interface.

* set register_for_poll as a private method

Although this method has a public facing pair in unregister_from_poll
it is currently never used outside of JournalReader. Making it private
to more clearly define the JournalReader interface.

* Make initialize_senders private

This method is not used anywhere outside of JournalReader and is made
private to more clearly define the usable JournalReader interface.

* make get_reader method private

The get_reader method is only used internally in the JournalReader
class and is made private to more clearly define the JournalReader
interface.