kafka sender: reduce try region and removed unecessary close
The `KafkaProducer` is the only point were an exception may be raised
in the `_init` method, if an exception is raised the assignment to
`self.kafka_producer` can not happen, so there is no need to do the
check on every iteration.

This also reduced moved the lines for the happy path out of the try
block, making the block contain only the lines that may raise an
exception.