Recreate reader if journal files is deleted
To allow this changes we always keep journal reader registered with
poller. For every even received we keep a lookup table of said events,
that should be processed at later stage.

There are 2 relevant types of events:
* journal.APPEND - when lines are added to the end of the file
* journal.INVALIDATE - log file rotated or removed

For APPEND event we need to read all the data in the log until it's
exhauisted

For INVALIDATE event data should be also read until the end and the
reader should be recreated to release fds of deleted files.

This fixes resource leak which happened if log file was first
renamed and then removed, which usually is the case when rotation
happens.