Add a secret filtering functionality to journalpump.
This change adds a secret filtering functionality. Config is at the reader level to allow one reader to redact secrets for multiple senders, but enable different behaviour between readers.
Provide a regex with groups, a redaction message, and define which groups are to be redacted in the case of a match. Multiple patterns can be added.

Fixes from review comments

Added a metric emit when filters hit a match.
Move secret filtering forward for reliable detection metrics
Added some configuration validation, and tests for the validation.