fix: redact sensitive information from indicies url when logging
`_indicies_url` for the `_EsOsLogSenderBase` class can contain sensitive
information, typically a password in the form of
`https://user:pass@host:port`. As a result, this information would be
logged should ES/OS not be available.

This commit redacts this using a standard pattern to remove this
password from being logged.