fix: tighten up tag sanitization
The tag sanitization introduced in
14a57be709cc9a5968918e0e25ea54f823c1fb5d was both too strict and too
loose. As per
https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels,
we need to make sure labels match the regexp `[a-zA-Z_:][a-zA-Z0-9_:]*`,
but `statsd` is more restrictive:
https://github.com/statsd/statsd/blob/9cf77d87855bcb69a8663135f59aa23825db9797/stats.js#L164-L172.

To solve this, lets have the tag names match strictly `[^\w_\-]+`.

For values, these can be typically any ASCII chars (as per Prometheus
above). However, the line protocol excludes the use of certain
characters, so lets ensure all characters that match `[\s\|=:]+` are
replaced.