journalpump: sort readers' fields and check on MESSAGE as last
At the moment, fields are sorted when serializing the config object.
Here fields are sorted when initializing the JournalPump object and also
the MESSAGE field is added as the last one to make the search operation
to first evaluate conditions on the rest of the fields and thus evaluate
the MESSAGE regexp only when necessary (for the sake of performance,
since the MESSAGE field will be longer than the others).