fix: resume from reader cursor when no senders configured
In cases where a reader was configured with no senders (eg. only
configured with searches), then in case of reloads/restarts the resume
cursor was assumed `None`, and the default 'initial_position' of 'head'
is fallen back to. This would end up causing all matching searchg to
resend metric events immediately.