senders: Skips messages larger than 256K
GCP quotas page for the logging v2(https://cloud.google.com/logging/quotas) specifies
that a LogEntry should have at maximum 256KB in size, that is not a hard limit, but
an estimation.
In anycase we should check if any of our LogEntry objects are bigger than the quota
allows and not send those logs. Instead we try to truncate the message to smaller size
and if that isn't enough we send a default log message.