tests: update rsyslog test to work
There are issues with python-systemd and the 'tail' behaviour,
this test will not work with it (no messages received after
starting journalpump). For example:

https://github.com/systemd/python-systemd/issues/146
https://github.com/systemd/python-systemd/issues/98
https://github.com/systemd/systemd/issues/17662

We also need to log via journal otherwise depending on the logging
configuration the test is prone to failure.

Finally, AppArmor on the ubuntu runner in CI blocks rsyslogd
by default so we just disable that profile here.