rsyslog: replace conf truncate_multiline by octet_counted_framing
The recently added config `truncate_multiline`(PR #180 and #181)
might raise wrong user expectation as it relies on a workaround
that is not supported by all syslog servers.

I propose to keep this dev but rename the config to
`octet_counted_framing` which is what it actually does, and
inverse the value. This should avoid confusion and prevent any
future misuse.